New columns

+++ app/src/main/java/com/example/muttersprache/MainActivity.kt
@@ -17,18 +17,19 @@ import java.util.Locale
 import java.util.Random
 import java.util.Timer
 import java.util.TimerTask
-import kotlin.math.log
 
+data class Quintuple<T1, T2, T3, T4, T5>(val first: T1, val second: T2, val third: T3, val fourth: T4, val fifth: T5)
 
 class MainActivity : ComponentActivity(), TextToSpeech.OnInitListener {
     private lateinit var textToSpeech: TextToSpeech
-    private lateinit var sentences: MutableList<Triple<String, String, String>>
+    private lateinit var sentences: MutableList<Quintuple<String, String, String, String, String>>
     private lateinit var timer: Timer
     private lateinit var listView: ListView
 
     private var speed: Float = 0.7f
     private var repeatTime:Long = 60000
     private var repeatSentence: String = ""
+    private var flagTimer: Boolean = false
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -42,7 +43,7 @@ class MainActivity : ComponentActivity(), TextToSpeech.OnInitListener {
 
         // Путь к файлу базы данных внутри внутреннего хранилища
         val databasePath = applicationContext.getDatabasePath(DATABASE_NAME).path
-
+        dbHelper.addColumnIfNeeded()
         // Проверяем, существует ли уже файл базы данных
         if (!File(databasePath).exists()) {
             // Открываем поток для копирования файла из assets во внутреннее хранилище
@@ -55,7 +56,6 @@ class MainActivity : ComponentActivity(), TextToSpeech.OnInitListener {
             while (inputStream.read(buffer).also { length = it } > 0) {
                 outputStream.write(buffer, 0, length)
             }
-
             // Закрываем потоки
             outputStream.flush()
             outputStream.close()
@@ -63,14 +63,17 @@ class MainActivity : ComponentActivity(), TextToSpeech.OnInitListener {
         }
 
         val database: SQLiteDatabase = SQLiteDatabase.openDatabase(databasePath, null, SQLiteDatabase.OPEN_READWRITE)
-        val cursor: Cursor = database.rawQuery("SELECT rowid, text, count FROM sentences", null)
+        val cursor: Cursor = database.rawQuery("SELECT rowid, text, count, translation, language FROM sentences", null)
+        dbHelper.newTable()
         sentences = mutableListOf()
         while (cursor.moveToNext()) {
             val value1: String = cursor.getString(0)
             val value2: String = cursor.getString(1)
             val value3: String = cursor.getString(2)
+            val value4: String = cursor.getString(3)
+            val value5: String = cursor.getString(4)
 
-            sentences.add(Triple(value1, value2, value3))
+            sentences.add(Quintuple(value1, value2, value3, value4, value5))
         }
         cursor.close()
         database.close()
@@ -83,12 +86,16 @@ class MainActivity : ComponentActivity(), TextToSpeech.OnInitListener {
 
         val editText: EditText = findViewById(R.id.editText)
 
+        val tranlsateText: EditText = findViewById(R.id.translateText)
+
         val addButton: Button = findViewById(R.id.addButton)
         addButton.setOnClickListener {
             val text = editText.text.toString()
+            val tranlsatetext = tranlsateText.text.toString()
             if (text.isNotEmpty()) {
-                dbHelper.addSentence(text, 0) // вставка данных, 0 - начальное значение счетчика
-                editText.text.clear() // очистка поля ввода
+                dbHelper.addSentence(text, tranlsatetext, 0, "ger") // вставка данных, 0 - начальное значение счетчика
+                editText.text.clear()
+                tranlsateText.text.clear()// очистка поля ввода
                 showSentences()
             }
         }
@@ -106,9 +113,11 @@ class MainActivity : ComponentActivity(), TextToSpeech.OnInitListener {
 
         val deleteButton: Button = findViewById(R.id.deleteIdButton)
         deleteButton.setOnClickListener {
-            val deleteID = deleteIDInput.text.toString().toInt()
-            dbHelper.deleteSentence(deleteID)
-            showSentences()
+            if(deleteIDInput.text.isNotEmpty()){
+                val deleteID = deleteIDInput.text.toString().toInt()
+                dbHelper.deleteSentence(deleteID)
+                showSentences()
+            }
         }
 
         val speedSeekBar: SeekBar = findViewById(R.id.speedSeekBar)
@@ -139,16 +148,22 @@ class MainActivity : ComponentActivity(), TextToSpeech.OnInitListener {
     }
 
     private fun startSpeech() {
-        timer = Timer()
-        timer.schedule(object : TimerTask() {
-            override fun run() {
-                speakRandomSentence()
-            }
-        }, 0, repeatTime) // Воспроизводить каждые 60 секунд
+        if(flagTimer == false) {
+            timer = Timer()
+            timer.schedule(object : TimerTask() {
+                override fun run() {
+                    speakRandomSentence()
+                }
+            }, 0, repeatTime)
+            flagTimer = true// Воспроизводить каждые 60 секунд
+        }
     }
 
     private fun stopSpeech() {
-        timer.cancel()
+        if(flagTimer) {
+            timer.cancel()
+            flagTimer = false
+        }
     }
 
     private fun speakRandomSentence(){
@@ -177,14 +192,16 @@ class MainActivity : ComponentActivity(), TextToSpeech.OnInitListener {
         val DATABASE_NAME = "new.db"
         val databasePath = applicationContext.getDatabasePath(DATABASE_NAME).path
         val database: SQLiteDatabase = SQLiteDatabase.openDatabase(databasePath, null, SQLiteDatabase.OPEN_READWRITE)
-        val cursor: Cursor = database.rawQuery("SELECT rowid, text, count FROM sentences ORDER BY count DESC", null)
+        val cursor: Cursor = database.rawQuery("SELECT rowid, text, count, translation, language FROM sentences ORDER BY count DESC", null)
         sentences = mutableListOf()
         while (cursor.moveToNext()) {
             val value1: String = cursor.getString(0)
             val value2: String = cursor.getString(1)
             val value3: String = cursor.getString(2)
+            val value4: String = cursor.getString(3)
+            val value5: String = cursor.getString(4)
 
-            sentences.add(Triple(value1, value2, value3))
+            sentences.add(Quintuple(value1, value2, value3, value4, value5))
         }
         cursor.close()
         database.close()
